requirements.txt
cryptography>=41.0.0
argon2-cffi>=23.1.0
pillow>=10.0.0
python-dotenv>=1.0.0
